8/9/2021_4:39PM
All places where the coded referred to animating were commented out. The scripts were not erased inside the Blender-Python API, but the checkbox in the GUI was removed, as well as the definitions for functions required to build the animation frames. Functions stored elsewhere. Wrapping up the version 2020.1.0 here. Animation will be a feature for future versions.