1/8/2025_7:12 to 9:76 AM
xyz_reader branch:
Starts with a fully functional program, capable of reading .com input files. At the moment, it is the Blender API which handles reading the .com input and getting the relevant coordinates. If I want to read .xyz files, I need to move the coordinate reading to my GUI instead, so that I can manage the two cases .com and .xyz and give the Blender API the same list of coordinates and atoms regardless of the extension of the input. Tried to populate a dictionary of dictionaries to store all the average single, double and triple bonds between the 118 elements. However, this is becoming longer than anticipated. Will leave this project to the future.