Sphynx documentation part 1
I'm using the Sphynx program to create an html file with all the classes, modules and functions for both the gui and the api.
I completed the task with the gui with no issue. Problem is, I need to run sphinx inside Blender to complete the documentation for the API. Will try that at another moment.