Before XYZ trajectory patch
Version of the program before the xyz trajectory reader idea. I plan to modify the Blender API portion of the program so that if one xyz file is read, and there is more than one set of coordinates for the same number and identity of atoms, the program must make an animation of them.