Tab system overhaul
In order to simplify user interactions, I wrapped each section of the GUI inside a tab. It currently works fine for static conversions. There is an issue with animations where although they're being converted as fbx by default, the file is renamed as glb in the end and there's errors there. Working on that.